server/requirements.txt
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
applaunchservices==0.2.1
appnope==0.1.0
appscript==1.0.1
argh==0.26.2
asn1crypto==1.3.0
astroid==2.3.3
astropy==4.0
async-timeout==4.0.3
atomicwrites==1.3.0
attrs==24.2.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.2
buildozer==1.2.0
cached-property==1.5.2
cachetools==5.3.1
caer==2.0.8
certifi==2024.8.30
cffi==1.14.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.4.0
Click==7.0
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
conda==4.8.2
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
coverage==7.2.7
cryptography==2.8
cycler==0.10.0
Cython==0.29.10
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.8
distributed==2.11.0
distro==1.9.0
dnspython==2.3.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastcache==1.1.0
ffpyplayer==4.3.1
filelock==3.12.2
flake8==3.7.9
Flask==1.1.1
Flask-Cors==5.0.0
fonttools==4.38.0
fsspec==0.6.2
future==0.18.2
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httpcore==0.17.3
httpx==0.24.1
hypothesis==5.5.4
identify==2.5.24
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==6.7.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==21.1.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kivy-garden.graph @ https://github.com/kivy-garden/graph/archive/master.zip
kivymd==0.104.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.5.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mypy==0.991
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
neo4j==5.27.0
networkx==2.4
nltk==3.4.5
nodeenv==1.8.0
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
olefile==0.46
openai==1.39.0
opencv-contrib-python==4.7.0.68
openpyxl==3.0.3
outcome==1.3.0.post0
packaging==24.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==1.0.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
platformdirs==4.0.0
playwright==1.35.0
pluggy==0.13.1
ply==3.11
pre-commit==2.21.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
proto-plus==1.22.3
protobuf==4.24.2
psutil==5.6.7
psycopg2-binary==2.9.9
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==4.0.1
pyee==9.0.4
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.5.2
PyInstaller==3.6
pylint==2.4.4
pymongo==4.7.3
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==3.0.1
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-cov==4.1.0
pytest-doctestplus==0.5.0
pytest-filter-subpackage==0.1.2
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-dotenv==0.21.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
redis==5.0.8
requests==2.31.0
rope==0.16.0
rsa==4.9
Rtree==0.9.3
ruamel_yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.7.3
seaborn==0.12.2
selenium==4.11.2
Send2Trash==1.5.0
sh==1.14.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
sniffio==1.3.1
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
statsmodels==0.13.5
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tdqm==0.0.1
terminado==0.8.3
testpath==0.4.4
tomli==2.0.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
trio==0.22.2
trio-websocket==0.11.1
typed-ast==1.5.4
typing_extensions==4.7.1
ujson==1.35
unicodecsv==0.14.1
urllib3==2.0.7
virtualenv==20.25.0
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wsproto==1.2.0
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==1.0.0
zipp==2.2.0
